refactor(generator): change ModelOperationType from enum to string literal type to improve type safety and readability
The ModelOperationType enum in the file `libs/generator/src/enums/model-operation-type.ts` was deleted as it was no longer being used. This change was made to remove unnecessary code and improve code cleanliness.

In the file `libs/generator/src/helpers/dmmf/internal-dmmf.ts`, the ModelOperationType enum was replaced with a string literal type to improve type safety and readability. The enum was imported from the deleted file, but since it was no longer needed, it was replaced with the string literal type directly in the code.

In the file `libs/generator/src/types/dmmf.d.ts`, the import statement for the ModelOperationType enum was removed as it was no longer being used.